Move message handler into chat handler
